requirements.txt
anyascii==0.3.1
autocuda==0.11
beautifulsoup4==4.11.1
bert-extractive-summarizer==0.10.1
blis==0.7.9
boostaug==2.2.5
Brotli==1.0.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
confection==0.0.3
contractions==0.1.72
cycler==0.11.0
cymem==2.0.7
Cython==0.29.28
dash==2.6.2
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1.tar.gz
et-xmlfile==1.1.0
filelock==3.8.0
findfile==1.7.9.8
Flask==2.2.2
Flask-Compress==1.13
fonttools==4.37.4
gdown==4.5.3
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.29
hdbscan==0.8.28
huggingface-hub==0.10.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
langcodes==3.3.0
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.5.3
metric-visualizer==0.6.3
murmurhash==1.0.9
natsort==8.2.0
networkx==2.8.7
nltk==3.7
numba==0.56.3
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pathy==0.6.2
Pillow==9.2.0
plotly==5.10.0
preshed==3.0.8
protobuf==3.20.2
PyABSA==1.16.18
pyahocorasick==1.4.4
pydantic==1.10.2
pynndescent==0.5.7
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-warmup==0.1.0
pytz==2022.5
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.0.1
textsearch==0.0.24
thinc==8.1.5
threadpoolctl==3.1.0
tikzplotlib==0.10.1
tokenizers==0.13.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.23.1
typer==0.4.2
typing_extensions==4.4.0
umap-learn==0.5.3
update-checker==0.18.0
urllib3==1.26.12
wasabi==0.10.1
webcolors==1.12
Werkzeug==2.2.2
wordcloud==1.8.2.2
